fix(android): repair the Capacitor APK pipeline after the monorepo migration
The Android release pipeline broke during the apps/web migration and was never
revalidated (mobile was deferred to "Milestone M"). Fixes so a published
release (or workflow_dispatch) produces a signed, installable APK again —
suitable for GitHub-release distribution (e.g. Obtainium), independent of Play.

- next.config.mjs: omit `rewrites`/`headers` under CAPACITOR_BUILD — both are
  unsupported by `output: export` and broke the static build. The hosted backend
  still serves them; the bundled shell calls it for /api/* so it needs neither.
- help/[slug]: split into a server page exporting generateStaticParams (all 13
  manual slugs from MANUALS) + a client child. A dynamic segment with no params
  can't be statically exported; also a free SSG win on the web build.
- capacitor.config.ts: moved to apps/web with `android.path: ../../android`, so
  `cap sync` (run from apps/web) discovers the @capacitor/* plugins from
  apps/web/package.json and writes into the root android/ project. Plugin/CLI
  discovery was silently broken (root package.json has no Capacitor deps).
- cap-build.js: read version from the root package.json (not apps/web's 0.0.0),
  write version.properties to the real android/app/, run `next build` in
  apps/web, and stash api/ + middleware OUTSIDE apps/web so Next's type-check
  (which covers all **/*.ts) doesn't choke on the moved routes' dangling imports.
- android-release.yml: run the web/cap steps from apps/web; drop the redundant
  version.properties step (cap-build owns it); always upload the signed APK as a
  run artifact (installable build from any run); let workflow_dispatch build the
  current branch (blank tag) for pre-release validation.
- docs + .gitignore updated.

Validated locally: `cap-build.js` static export is green (29 pages, /help/[slug]
SSG with all slugs), stash restores cleanly, web typecheck + lint pass. The
Gradle/sign/APK half is validated via CI (all four ANDROID_KEYSTORE_* secrets
are already set; no PLAY secret, so the Play-upload step auto-skips).
